remove size_t from parser

+++ include/mruby/compile.h
@@ -91,12 +91,12 @@ void mrb_parser_parse(struct mrb_parser_state*);
 /* utility functions */
 struct mrb_parser_state* mrb_parse_file(mrb_state*,FILE*);
 struct mrb_parser_state* mrb_parse_string(mrb_state*,const char*);
-struct mrb_parser_state* mrb_parse_nstring(mrb_state*,const char*,size_t);
+struct mrb_parser_state* mrb_parse_nstring(mrb_state*,const char*,int);
 int mrb_generate_code(mrb_state*, mrb_ast_node*);
 
 int mrb_compile_file(mrb_state*,FILE*);
 int mrb_compile_string(mrb_state*,char*);
-int mrb_compile_nstring(mrb_state*,char*,size_t);
+int mrb_compile_nstring(mrb_state*,char*,int);
 
 #if defined(__cplusplus)
 }  /* extern "C" { */
